requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crispy-bootstrap5==2025.6
cron_descriptor==2.0.6
dj-database-url==3.0.1
Django==5.2.7
django-celery-beat==2.8.1
django-crispy-forms==2.4
django-timezone-field==7.1
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
numpy==2.3.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
